fix: #1 기본 레이아웃 구조 및 오류 수정
