--- src/components/common/Button/ProductDeleteButton.jsx
@@ -1,5 +1,0 @@
-const ProductDeleteButton = () => {
-    return (<></>);
-}
-
-export default ProductDeleteButton;
