+++ src/pages/ProductRegisterPage.jsx
@@ -1,6 +1,5 @@
 import ProductRegisterForm from "../components/Product/ProductRegisterForm";
-import HeaderBack from "../components/common/Header/HeaderBack";
-import Navbar from "../components/common/Navbar";
+
 const ProductRegisterPage = () => {
